# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Mining Protocol for Inscription Miners Based on the BTC Ordinals Protocol",
        "title": "BTC Ordinals Mining Protocol",
        "termsOfService": "https://github.com/KingKaiBTC",
        "contact": {},
        "version": "1.01"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/mrc20/getrecords": {
            "get": {
                "description": "Retrieves a list of records based on address, record type, page number, page size, and order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "record"
                ],
                "summary": "Retrieve records from the database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address identifier",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record type identifier",
                        "name": "rectype",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of records per page",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Order of record retrieval (true for ascending, false for descending)",
                        "name": "ascend",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of records",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Error message if retrieval fails",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mrc20/hookevents": {
            "post": {
                "description": "Parses and processes events related to OrdHook.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrdHook"
                ],
                "summary": "Process OrdHook events",
                "parameters": [
                    {
                        "description": "Event Payload",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rpc.OrdHookEvent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A message confirming successful processing",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error message in case of failure to process the event",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/mrc20/postrecord": {
            "post": {
                "description": "Writes a new record with an incrementing index to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "record"
                ],
                "summary": "Write a record to the database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address identifier",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record type identifier",
                        "name": "rectype",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Message content to be stored",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message with code and message text",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error message with code and message text",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "rpc.InscriptionNumber": {
            "type": "object",
            "properties": {
                "classic": {
                    "type": "integer"
                },
                "jubilee": {
                    "type": "integer"
                }
            }
        },
        "rpc.OrdHookBlock": {
            "type": "object",
            "properties": {
                "block_identifier": {
                    "$ref": "#/definitions/rpc.OrdHookBlockIdentifier"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "parent_block_identifier": {
                    "$ref": "#/definitions/rpc.OrdHookBlockIdentifier"
                },
                "timestamp": {
                    "type": "integer"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rpc.OrdHookTransaction"
                    }
                }
            }
        },
        "rpc.OrdHookBlockIdentifier": {
            "type": "object",
            "properties": {
                "hash": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                }
            }
        },
        "rpc.OrdHookChain": {
            "type": "object",
            "properties": {
                "is_streaming_blocks": {
                    "type": "boolean"
                },
                "predicate": {
                    "$ref": "#/definitions/rpc.OrdHookPredicate"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "rpc.OrdHookDestination": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "rpc.OrdHookEvent": {
            "type": "object",
            "properties": {
                "apply": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rpc.OrdHookBlock"
                    }
                },
                "chainhook": {
                    "$ref": "#/definitions/rpc.OrdHookChain"
                },
                "rollback": {
                    "description": "Assuming rollback has a similar structure",
                    "type": "array",
                    "items": {}
                }
            }
        },
        "rpc.OrdHookInscriptionRevealed": {
            "type": "object",
            "properties": {
                "content_bytes": {
                    "type": "string"
                },
                "content_length": {
                    "type": "integer"
                },
                "content_type": {
                    "type": "string"
                },
                "curse_type": {
                    "description": "Assuming curse_type can be null, hence using a pointer",
                    "type": "string"
                },
                "inscriber_address": {
                    "type": "string"
                },
                "inscription_fee": {
                    "type": "integer"
                },
                "inscription_id": {
                    "type": "string"
                },
                "inscription_input_index": {
                    "type": "integer"
                },
                "inscription_number": {
                    "$ref": "#/definitions/rpc.InscriptionNumber"
                },
                "inscription_output_value": {
                    "type": "integer"
                },
                "ordinal_block_height": {
                    "type": "integer"
                },
                "ordinal_number": {
                    "type": "integer"
                },
                "ordinal_offset": {
                    "type": "integer"
                },
                "satpoint_post_inscription": {
                    "type": "string"
                },
                "transfers_pre_inscription": {
                    "type": "integer"
                },
                "tx_index": {
                    "type": "integer"
                }
            }
        },
        "rpc.OrdHookInscriptionTransferred": {
            "type": "object",
            "properties": {
                "destination": {
                    "$ref": "#/definitions/rpc.OrdHookDestination"
                },
                "inscription_id": {
                    "type": "string"
                },
                "post_transfer_output_value": {
                    "type": "integer"
                },
                "satpoint_post_transfer": {
                    "type": "string"
                },
                "satpoint_pre_transfer": {
                    "type": "string"
                },
                "tx_index": {
                    "type": "integer"
                }
            }
        },
        "rpc.OrdHookOrdinalOperation": {
            "type": "object",
            "properties": {
                "inscription_revealed": {
                    "$ref": "#/definitions/rpc.OrdHookInscriptionRevealed"
                },
                "inscription_transferred": {
                    "$ref": "#/definitions/rpc.OrdHookInscriptionTransferred"
                }
            }
        },
        "rpc.OrdHookPredicate": {
            "type": "object",
            "properties": {
                "operation": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                }
            }
        },
        "rpc.OrdHookTransaction": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/rpc.OrdHookTransactionMetadata"
                },
                "operations": {
                    "type": "array",
                    "items": {}
                },
                "transaction_identifier": {
                    "$ref": "#/definitions/rpc.OrdHookBlockIdentifier"
                }
            }
        },
        "rpc.OrdHookTransactionMetadata": {
            "type": "object",
            "properties": {
                "ordinal_operations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rpc.OrdHookOrdinalOperation"
                    }
                },
                "proof": {}
            }
        },
...
    }
}